% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkListFitparam}
\alias{checkListFitparam}
\title{Check a list of fitparam objects}
\usage{
checkListFitparam(list.fitparam)
}
\arguments{
\item{list.fitparam}{named list of \code{\link{fitparam}} objects.}
}
\description{
Check a list of fitparam objects
}

