% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compositeLogPrior}
\alias{compositeLogPrior}
\title{Compute a composite log-prior}
\usage{
compositeLogPrior(list.fitparam)
}
\arguments{
\item{list.fitparam}{A list of \code{fitparam} objects. If they have the
\code{prior} argument set,  they will have their prior
distribution (given by \code{distribution} and \code{parameters}
within the \code{prior} of the \code{fitparam} object) evaluated
at its \code{value}}
}
\value{
Composite log-prior probability.
}
\description{
Caculates log of the prior probability at a given (set of) parameter value(s),
by composing the univariate prior distribution of each theta.
}

