% Generated by roxygen2 (4.0.1): do not edit by hand
\name{computeDistanceABC}
\alias{computeDistanceABC}
\title{Compute the distance to the data for ABC}
\usage{
computeDistanceABC(theta, fitmodel)
}
\arguments{
\item{theta}{named vector of model parameters. Names must correspond to those of \code{names(fitmodel$theta)}.}

\item{fitmodel}{\code{\link{fitmodel}} object.}
}
\value{
numeric value of the log-likelihood
}
\description{
Compute the distance (using \code{fitmodel$distance.ABC}) between the observed time series and a simulated time series obtained by running the model with parameters \code{theta}.
}

