% Generated by roxygen2 (4.0.1): do not edit by hand
\name{distanceOscillation}
\alias{distanceOscillation}
\title{Distance weighted by number of oscillations}
\usage{
distanceOscillation(x, y)
}
\arguments{
\item{x,y}{numeric vectors of the same length.}
}
\description{
This positive distance is the mean squared differences between \code{x} and the \code{y}, divided by the square of the number of times the \code{x} oscillates around the \code{y} (see note below for illustration).
}
\note{
To illustrate this distance, suppose we observed a time series \code{y = c(1,3,5,7,5,3,1)} and we have two simulated time series \code{x1 = (3,5,7,9,7,5,3)} and \code{x2 = (3,5,3,5,7,5,3)}; \code{x1} is consistently above \code{y} and \code{x2} oscillates around \code{y}. While the squared differences are the same, we obtain \eqn{d(y, x1) = 4} and \eqn{d(y, x2) = 1.3}.
}

