% Generated by roxygen2 (4.0.1): do not edit by hand
\name{export2Tracer}
\alias{export2Tracer}
\title{Export trace in Tracer format}
\usage{
export2Tracer(trace, file)
}
\arguments{
\item{trace}{a \code{data.frame} with one column per estimated parameter, as returned by \code{\link{burnAndThin}}}

\item{file}{either a character string naming a file or a \link{connection}
    open for writing.  \code{""} indicates output to the console.}
}
\description{
Print \code{trace} in a \code{file} that can be read by the software Tracer.
}
\note{
Tracer is a program for analysing the trace files generated by Bayesian MCMC runs. It can be dowloaded at \url{http://tree.bio.ed.ac.uk/software/tracer}.
}
\seealso{
burnAndThin
}

