% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitparam}
\alias{fitparam}
\title{Constructor of fitparam object}
\usage{
fitparam(name = NULL, value = NULL, support = c(-Inf, Inf),
  sd.proposal = 0, prior = list(distribution = NULL, parameters = NULL))
}
\arguments{
\item{name}{character, name of the parameter.}

\item{value}{numeric, value of the parameter.}

\item{support}{numeric vector, range of the support of the parameter.}

\item{sd.proposal}{numeric, standard deviation of the proposal gaussian kernel.}

\item{prior}{list of 2 elements:
\itemize{
    \item \code{distribution} character, name of density function for the prior distribution. IMPORTANT: it must be the name of recognized \R function (ex: "dnorm", "dunif").
    \item \code{parameters} named vector of parameters for the prior distribution, as specified in the documentation of the \R function. For instance, the \code{mean} and \code{sd} of a normal distribution, or the \code{min} and \code{max} of a uniform distribution.
}}
}
\value{
a \code{fitparam} object
}
\description{
A \code{fitparam} object contains all the information necessary to fit a model parameter.
}
\examples{
\dontrun{
myParam <- fitparam(name="R0",value=3,support=c(0,Inf),sd.proposal=1,prior=list(distribution="uniform",parameters=c(min=1,max=100)))
}
}
\seealso{
getParameterValues, setParameterValues
}

