% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getParameterValues}
\alias{getParameterValues}
\title{Get parameter values from a list of fitparam objects}
\usage{
getParameterValues(list.fitparam)
}
\arguments{
\item{list.fitparam}{named list of \code{\link{fitparam}} objects.}
}
\value{
a named vector of parameter values
}
\description{
If \code{list.fitparam} is a list of \code{\link{fitparam}} objects, the value of the theta will be returned as a named vector, possibly restricting to estimated theta only.
}
\examples{
\dontrun{
list.fitparam <- list(fitparam(name="A",value=1,support=c(0,10),prior=list(distribution="uniform",theta=c(min=0,max=10))),fitparam(name="B",value=2),fitparam(name="C",value=2))
getParameterValues(list.fitparam)
}
}

