% Generated by roxygen2 (4.0.1): do not edit by hand
\name{marginalLogLikelihoodDeterministic}
\alias{marginalLogLikelihoodDeterministic}
\title{Marginal log-likelihood for a deterministic model}
\usage{
marginalLogLikelihoodDeterministic(theta, fitmodel)
}
\arguments{
\item{theta}{named vector of model parameters. Names must correspond to those of \code{names(fitmodel$theta)}.}

\item{fitmodel}{\code{\link{fitmodel}} object.}
}
\value{
numeric value of the log-likelihood
}
\description{
Compute the marginal log-likelihood of \code{theta} for a deterministic model defined in a \code{\link{fitmodel}} object.
}

