% Generated by roxygen2 (4.0.1): do not edit by hand
\name{marginalLogLikelihoodStochastic}
\alias{marginalLogLikelihoodStochastic}
\title{Marginal log-likelihood for a stochastic model}
\usage{
marginalLogLikelihoodStochastic(theta, fitmodel, n.particles, n.cores = 1)
}
\arguments{
\item{theta}{named vector of model parameters. Names must correspond to those of \code{names(fitmodel$theta)}.}

\item{fitmodel}{\code{\link{fitmodel}} object.}

\item{n.particles}{number of particles}

\item{n.cores}{number of cores on which propogation of the particles is parallelised. By default no parallelisation (\code{n.cores=1}). If \code{NULL}, set to the value returned by \code{\link[parallel]{detectCores}}.}
}
\value{
Monte-Carlo estimate of the marginal log-likelihood of \code{theta}
}
\description{
Compute a Monte-Carlo estimate of the log-likelihood of \code{theta} for a stochastic model defined in a \code{\link{fitmodel}} object, using \code{\link{bootstrapParticleFilter}}
}
\seealso{
bootstrapParticleFilter
}

