% Generated by roxygen2 (4.0.1): do not edit by hand
\name{measles}
\alias{measles}
\title{Time-series of a measles outbreak}
\format{A data frame with 37 rows and 2 variables}
\description{
A dataset containing the weekly incidence recorded during a recent outbreak of measles in Europe
}
\details{
\itemize{
  \item \code{time} week of the record
  \item \code{Inc} weekly recorded incidence of measles
}
}

