% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotModelTraj}
\alias{plotModelTraj}
\title{Plot one or more model trajectories}
\usage{
plotModelTraj(fitmodel, traj, state.variables = NULL, alpha = 1,
  plot = TRUE)
}
\arguments{
\item{fitmodel}{a \code{\link{fitmodel}} object.}

\item{traj}{data.frame, output of \code{fitmodel$simulate.model} or \code{simulateModelReplicates}.}

\item{state.variables}{subset of state.variables to plot. If \code{NULL} (default) all state variables are plotted.}

\item{alpha}{transparency of the trajectories (between 0 and 1).}

\item{plot}{if \code{TRUE} the plot is displayed, and returned otherwise.}
}
\description{
This function use faceting to plot all state.variables trajectories. Convenient to see results of several simulations.
}
\seealso{
simulateModelReplicates
}

