% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotPosteriorFit}
\alias{plotPosteriorFit}
\title{Plot MCMC posterior fit}
\usage{
plotPosteriorFit(trace, fitmodel, posterior.median = FALSE, summary = FALSE,
  sample.size = 100, alpha = min(1, 10/sample.size), plot = TRUE)
}
\arguments{
\item{posterior.median}{logical, if \code{TRUE} use the median of the posterior distribution.}

\item{summary}{logical, if \code{TRUE} trajectories are summarised by their mean, median, 50\% and 95\% quantile distributions. Otheriwse, the trajectories are ploted.}

\item{sample.size}{number of replicated simulations (if \code{posterior.median=TRUE}) or number of theta sampled from posterior distribution (if \code{posterior.median=TRUE}).}

\item{trace}{a \code{data.frame} with one column per estimated parameter, as returned by \code{\link{burnAndThin}}}

\item{fitmodel}{a \code{\link{fitmodel}} object.}

\item{alpha}{transparency of the trajectories (between 0 and 1).}

\item{plot}{if \code{TRUE} the plot is displayed, and returned otherwise.}
}
\value{
If \code{plot==FALSE}, a list of 2 elements is returned:
\itemize{
   \item \code{posterior.traj} a \code{data.frame} with the trajectories (and observations) sampled from the posterior distribution.
   \item \code{plot} the plot of the fit displayed.
}
}
\description{
Plot posterior distribution of observation generated under model's posterior parameter distribution.
}

