% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotSMC}
\alias{plotSMC}
\title{Plot result of SMC}
\usage{
plotSMC(smc, fitmodel, alpha = 1, plot = TRUE)
}
\arguments{
\item{smc}{output of \code{\link{bootstrapParticleFilter}}}

\item{fitmodel}{a \code{\link{fitmodel}} object.}

\item{alpha}{transparency of the trajectories (between 0 and 1).}

\item{plot}{if \code{TRUE} the plot is displayed, and returned otherwise.}
}
\description{
Plot the observation generated by the filtered trajectories together with the data.
}
\seealso{
bootstrapParticleFilter
}

