% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotThetaFit}
\alias{plotThetaFit}
\title{Plot fit of model to data}
\usage{
plotThetaFit(theta, fitmodel, n.replicates = 1, alpha = min(1,
  10/n.replicates), plot = TRUE)
}
\arguments{
\item{n.replicates}{numeric, number of replicated simulations.}

\item{theta}{named vector of model parameters. Names must correspond to those of \code{names(fitmodel$theta)}.}

\item{fitmodel}{\code{\link{fitmodel}} object.}

\item{alpha}{transparency of the trajectories (between 0 and 1).}

\item{plot}{if \code{TRUE} the plot is displayed, and returned otherwise.}
}
\value{
if \code{plot==FALSE}, a list of 2 elements is returned:
\itemize{
    \item \code{simulations} \code{data.frame} of \code{n.replicates} simulated observations.
    \item \code{plot} the plot of the fit.
}
}
\description{
Simulate the model under \code{theta}, generate observation and plot against data. Since simulation and observation processes can be stochastic, \code{n.replicates} can be plotted.
}

