% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotTrace}
\alias{plotTrace}
\title{Plot MCMC trace}
\usage{
plotTrace(trace, estimated.only = FALSE)
}
\arguments{
\item{trace}{a \code{data.frame} with one column per estimated parameter, as returned by \code{\link{burnAndThin}}}

\item{estimated.only}{logical, if \code{TRUE} only estimated parameters are displayed.}
}
\description{
Plot the traces of all estimated variables.
}
\seealso{
burnAndThin
}

