% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SEIT2L_distanceOscillation}
\alias{SEIT2L_distanceOscillation}
\title{ABC distance with oscillations}
\usage{
SEIT2L_distanceOscillation(model.traj.obs, data)
}
\arguments{
\item{model.traj.obs}{\code{data.frame} of simulated trajectory with observation, as returned by \code{\link{SEIT2L_generateObservation}}.}

\item{data}{\code{data.frame} of times and observations. Must have two columns: \code{time} and \code{Inc}.}
}
\description{
This positive distance is the mean squared differences between the simulation and the observation, divided by the square of the number of times the simulation oscillates around the observation.
}
\examples{
\dontrun{
# Suppose we observed a time series:
data <- data.frame(time=1:7,Inc=c(1,3,5,7,5,3,1))
# and we have two simulated time series:
traj1 <- data.frame(time=1:7,observation=c(3,5,7,9,7,5,3))
traj2 <- data.frame(time=1:7,observation=c(3,5,3,5,7,5,3))
# traj1 is consistently above data and traj2 oscillates around data:
plot(data$time,data$Inc,t='l',ylim=c(0,10))
lines(traj1$time,traj1$observation,col="red")
lines(traj2$time,traj2$observation,col="blue")
# While the squared differences are the same, we obtain a smaller distance for traj2:
d1 <- SEIT2L_distanceOscillation(traj1,data)
# d1 = 4
d2 <- SEIT2L_distanceOscillation(traj2,data)
# d2 = 1.3
}
}
\seealso{
distanceOscillation
}

