% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SEIT2L_generateObservation}
\alias{SEIT2L_generateObservation}
\title{Generate an observed incidence time series}
\usage{
SEIT2L_generateObservation(model.traj, theta)
}
\arguments{
\item{model.traj}{simulated trajectory, as returned by \code{\link{SEIT2L_simulateDeterministic}} or \code{\link{SEIT2L_simulateStochastic}}.}

\item{theta}{named vector of parameters for the SEIT2L model.}
}
\value{
the \code{model.traj} data.frame with an additional variable: "observation".
}
\description{
Generate a daily incidence time serie under a Poisson observation process.
}
\seealso{
SEIT2L_likelihood
}

