% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SEIT2L_initialiseState}
\alias{SEIT2L_initialiseState}
\title{Initialise SEIT2L model}
\usage{
SEIT2L_initialiseState(theta)
}
\arguments{
\item{theta}{named vector of parameters for the SEIT2L model.}
}
\value{
a named vector of initial conditions for the SEIT2L model.
}
\description{
Set the initial value of the state space of the SEIT2L model for a given set of theta.
}
\note{
This function throws an error if the sum of all compartments is higher than the population size N.
}
\seealso{
fitmodel, generateSEIT2LmodelTDC
}

