% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SEIT2L_logPrior}
\alias{SEIT2L_logPrior}
\title{Log-prior for SEIT2L model}
\usage{
SEIT2L_logPrior(list.fitparam)
}
\arguments{
\item{list.fitparam}{list of \code{fitparam} objects.}
}
\value{
the value of the log-prior.
}
\description{
Compute the log of the prior density distribution for the parameter set \code{theta}, by composing the univariate priors of all estimated theta.
}
\note{
This function returns \code{-Inf} if the sum of all compartments is higher than the population size N.
}
\seealso{
fitmodel, compositeLogPrior
}

