% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SEIT2L_simulateDeterministic}
\alias{SEIT2L_simulateDeterministic}
\title{Deterministic simulation of SEIT2L model}
\usage{
SEIT2L_simulateDeterministic(theta, state.init, times)
}
\arguments{
\item{state.init}{a named vector with the initial state values for the SEIT2L model.}

\item{times}{time sequence for which output is wanted; the first value of \code{times} must be the initial time.}

\item{theta}{named vector of parameters for the SEIT2L model.}
}
\value{
a \code{data.fame} containing the values of the state variables (columns) at each observation times (rows).
}
\description{
Solves the system of ordinary differential equations for the SEIT2L model using the \code{\link[deSolve]{ode}} function.
}

