% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SEIT2L_simulateStochastic}
\alias{SEIT2L_simulateStochastic}
\title{Stochastic simulation of SEIT2L model}
\usage{
SEIT2L_simulateStochastic(theta, state.init, times)
}
\arguments{
\item{theta}{named vector of parameters for the SEIT2L model.}

\item{state.init}{a named vector with the initial state values for the SEIT2L model.}

\item{times}{time sequence for which output is wanted; the first value of \code{times} must be the initial time.}
}
\value{
a \code{data.fame} containing the values of the state variables (columns) at each observation times (rows).
}
\description{
Simulate realisation of the stochastic version of the SEIT2L model using the \code{simulateModelStochastic} function.
}
\seealso{
fitmodel, simulateModelStochastic
}

