% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SEITL_createModelTdC}
\alias{SEITL_createModelTdC}
\title{Create the SEITL model as a fitmodel object}
\usage{
SEITL_createModelTdC(deterministic = TRUE, verbose = TRUE)
}
\arguments{
\item{deterministic}{if \code{TRUE} then the model will be deterministic, and stochastic otherwise.}

\item{verbose}{if \code{TRUE}, print details of the test performed to check validity of the arguments}
}
\value{
a fitmodel object
}
\description{
This function returns a fitmodel object contaning the (either deterministic or stochastic) SEITL model, to be fitted to the \code{\link{FluTdC1971}} dataset.
}

