% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SEITL_initialiseState}
\alias{SEITL_initialiseState}
\title{Initialise SEITL model}
\usage{
SEITL_initialiseState(theta)
}
\arguments{
\item{theta}{named vector of parameters for the SEITL model.}
}
\value{
a named vector of initial conditions for the SEITL model.
}
\description{
Set the initial value of the state space of the SEITL model for a given set of theta.
}
\note{
This function throws an error if the sum of all compartments is higher than the population size N.
}
\seealso{
fitmodel, generateSEITLmodelTDC
}

