% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SEITL_logLikelihood}
\alias{SEITL_logLikelihood}
\title{Likelihood of the data for SEITL model}
\usage{
SEITL_logLikelihood(data, theta, model.traj)
}
\arguments{
\item{data}{subset of the \code{\link{FluTdC1971}} dataset that contains the observations corresponding to \code{model.traj}.}

\item{model.traj}{simulated trajectory, as returned by \code{\link{SEITL_simulateDeterministic}} or \code{\link{SEITL_simulateStochastic}}.}

\item{theta}{named vector of parameters for the SEITL model.}
}
\value{
the log-likelihood value.
}
\description{
Computes the log-likelihood of a subset of the data for a fixed trajectory and under a poisson observation process.
}
\note{
This function can be used to compute the likelihood of the data given a deterministic trajectory or the weight of a SMC particle at a given observation time.
}
\seealso{
SEITL_generateObservation
}

