% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SEITL_simulateStochastic}
\alias{SEITL_simulateStochastic}
\title{Stochastic simulation of SEITL model}
\usage{
SEITL_simulateStochastic(theta, state.init, times)
}
\arguments{
\item{theta}{named vector of parameters for the SEITL model.}

\item{state.init}{a named vector with the initial state values for the SEITL model.}

\item{times}{time sequence for which output is wanted; the first value of \code{times} must be the initial time.}
}
\value{
a \code{data.fame} containing the values of the state variables (columns) at each observation times (rows).
}
\description{
Simulate realisation of the stochastic version of the SEITL model using the \code{simulateModelStochastic} function.
}
\seealso{
fitmodel, simulateModelStochastic
}

