% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setParameterValues}
\alias{setParameterValues}
\title{Replace specified parameter values with new values, in a list of fitparam objects.}
\usage{
setParameterValues(list.fitparam, new.values)
}
\arguments{
\item{list.fitparam}{list of \code{\link{fitparam}} objects.}

\item{new.values}{named vector containing the new parameter values; names must match those of \code{list.fitparam}.}
}
\value{
a named list of \code{\link{fitparam}} objects.
}
\description{
If \code{list.fitparam} is a list of \code{\link{fitparam}} objects, the value of the theta will be replaced by those of \code{new.values}, by matching parameter names with names.
}
\examples{
\dontrun{
list.fitparam <- list(fitparam(name="A",value=1),fitparam(name="B",value=2),fitparam(name="C",value=2))
list.fitparam <- setParameterValues(list.fitparam,new.value=c("A"=10,"B"=5))
print(list.fitparam)
}
}
\seealso{
fitparam
}

