% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simulateModelReplicates}
\alias{simulateModelReplicates}
\title{Simulate several replicate of the model}
\usage{
simulateModelReplicates(fitmodel, theta = NULL, times, n,
  observation = FALSE)
}
\arguments{
\item{fitmodel}{a \code{\link{fitmodel}} object.}

\item{theta}{named vector of parameters. If \code{NULL}, use \code{fitmodel$theta} (default).}

\item{times}{vector of times at which you want to observe the states of the model.}

\item{n}{number of replicated simulations.}

\item{observation}{logical, if \code{TRUE} simulated obseravtion are generated by \code{fitmodel$generate.observation}.}
}
\value{
a data.frame of dimension \code{[nxlength(times)]x[length(state.init)+2]} with column names equal to \code{c("replicate","time",names(state.init))}.
}
\description{
Simulate several replicate of a fitmodel using its function simulate.model
}

