% Generated by roxygen2 (4.0.1): do not edit by hand
\name{targetPosterior}
\alias{targetPosterior}
\title{Target posterior distribution for a fitmodel}
\usage{
targetPosterior(theta, log.prior, log.prior.args = list(),
  marginal.log.likelihood, marginal.log.likelihood.args = list())
}
\arguments{
\item{theta}{named vector of estimated theta}

\item{log.prior}{\R-function to compute the log-prior of \code{theta}, as returned by \code{\link{fitmodel}}}

\item{log.prior.args}{list of arguments passed to \code{log.prior}}

\item{marginal.log.likelihood}{\R-function to compute the marginal log-likelihood of \code{theta}}

\item{marginal.log.likelihood.args}{list of arguments passed to \code{marginal.log.likelihood}}
}
\value{
a list of two elements
\itemize{
	\item \code{log.dist} numeric, logged value of the posterior distribution evaluated at \code{theta}
	\item \code{trace} named vector with trace information (theta, log.prior, marginal.log.likelihood, log.posterior)
}
}
\description{
This function evaluates the posterior distribution at \code{theta} and returns the result in a suitable format for \code{\link{mcmcMH}}.
}

