% Generated by roxygen2 (4.0.1): do not edit by hand
\name{targetPosteriorABC}
\alias{targetPosteriorABC}
\title{Target ABC posterior distribution for a fitmodel}
\usage{
targetPosteriorABC(theta, fitmodel, epsilon)
}
\arguments{
\item{epsilon}{numeric vector, ABC tolerances for distances between data and simulations.}

\item{theta}{named vector of estimated theta}

\item{fitmodel}{\code{\link{fitmodel}} object.}
}
\value{
a list of two elements
\itemize{
	\item \code{log.dist} numeric, logged value of the ABC posterior distribution evaluated at \code{theta}
	\item \code{trace} named vector with trace information (theta, log.prior, distance.ABC, log.posterior)
}
}
\description{
This function evaluates the ABC posterior distribution at \code{theta} and returns the result in a suitable format for \code{\link{mcmcMH}}.
}
\seealso{
computeDistanceABC
}

